devex-1197: fix paths after file move
